Util function for JsonObject
