+++ settings.gradle.kts
@@ -17,11 +17,13 @@ rootProject.name = "knotx-dropwizard-metrics"
 
 include("knotx-dropwizard-core")
 include("knotx-dropwizard-api")
+include("knotx-dropwizard-reporter-common")
 include("knotx-dropwizard-reporter-console")
 include("knotx-dropwizard-reporter-graphite")
 include("knotx-dropwizard-reporter-logger")
 project(":knotx-dropwizard-core").projectDir = file("core")
 project(":knotx-dropwizard-api").projectDir = file("api")
+project(":knotx-dropwizard-reporter-common").projectDir = file("reporter/common")
 project(":knotx-dropwizard-reporter-console").projectDir = file("reporter/console")
 project(":knotx-dropwizard-reporter-graphite").projectDir = file("reporter/graphite")
 project(":knotx-dropwizard-reporter-logger").projectDir = file("reporter/logger")
